requirements.txt
openai
guidance
python-dotenv
tqdm
click
langchain
pandas
